Split PlayersRow
