+++ src/GameRoom/index.tsx
@@ -8,12 +8,9 @@ import {
   useSetInnerState
 } from 'react-inner-store'
 import { toast } from 'react-hot-toast'
-import { differenceInMilliseconds } from 'date-fns'
 
-import { pct, formatDurationMs, useEvery } from '../utils'
 import {
   Game,
-  Player,
   Dice,
   useCompleteTurn,
   useQueryGame,
@@ -26,6 +23,7 @@ import * as Icon from '../Icon'
 import { DieEvent, DieNumber } from '../domain'
 
 import * as DieRow from './DieRow'
+import { PlayersRow } from './PlayersRow'
 
 export abstract class State {
   abstract readonly isMutating: InnerStore<boolean>
@@ -64,109 +62,6 @@ export abstract class State {
     return { whiteDie, redDie, eventDie }
   }
 }
-
-const ViewPlayerTile: React.VFC<{
-  isCurrentPlayer: boolean
-  player: Player
-}> = React.memo(({ isCurrentPlayer, player }) => {
-  return (
-    <div
-      className={cx(
-        'flex-1 flex justify-center h-6 transition-[font-size] duration-300',
-        isCurrentPlayer ? 'text-3xl' : 'text-5xl'
-      )}
-    >
-      <Icon.User style={{ color: player.color.hex }} />
-    </div>
-  )
-})
-
-const ViewCurrentPlayerCaret: React.VFC<{
-  isGamePaused: boolean
-  currentTurnDurationMs: number
-  currentTurnDurationSince: Date
-  playersCount: number
-  currentPlayerIndex: number
-}> = React.memo(
-  ({
-    isGamePaused,
-    currentTurnDurationMs,
-    currentTurnDurationSince,
-    playersCount,
-    currentPlayerIndex
-  }) => {
-    const fraction = 100 / playersCount
-    const currentTurnDuration = useEvery(
-      now => {
-        if (isGamePaused) {
-          return formatDurationMs(currentTurnDurationMs)
-        }
-
-        const diffMs = differenceInMilliseconds(now, currentTurnDurationSince)
-
-        return formatDurationMs(currentTurnDurationMs + diffMs)
-      },
-      {
-        interval: 60,
-        skip: isGamePaused
-      }
-    )
-
-    return (
-      <div
-        className="absolute inset-0 transition-transform ease-out duration-300"
-        style={{
-          transform: `translateX(${pct(fraction * currentPlayerIndex)})`
-        }}
-      >
-        <div
-          className="flex flex-col h-full justify-end ring-inset ring-2 ring-gray-200 rounded overflow-hidden"
-          style={{
-            width: pct(fraction)
-          }}
-        >
-          <span className="p-0.5 font-mono text-xs text-center text-gray-500 bg-gray-200">
-            {currentTurnDuration}
-          </span>
-        </div>
-      </div>
-    )
-  }
-)
-
-const ViewGamePlayers: React.VFC<{
-  game: Game
-}> = React.memo(({ game }) => {
-  const prevTurn = game.turns[0]
-  const currentPlayerIndex = React.useMemo(() => {
-    return Math.max(
-      0,
-      game.players.findIndex(player => {
-        return player.id === prevTurn?.player.nextPlayerId
-      })
-    )
-  }, [game.players, prevTurn])
-
-  return (
-    <div className="flex relative pt-2 pb-8">
-      <ViewCurrentPlayerCaret
-        isGamePaused={game.isPaused}
-        currentTurnDurationMs={game.currentTurnDurationMs}
-        currentTurnDurationSince={game.currentTurnDurationSince}
-        playersCount={game.players.length}
-        currentPlayerIndex={currentPlayerIndex}
-      />
-
-      {game.players.map((player, index) => (
-        <ViewPlayerTile
-          key={player.id}
-          isCurrentPlayer={index === currentPlayerIndex}
-          player={player}
-        />
-      ))}
-    </div>
-  )
-})
 
 const ViewCompleteTurnButton: React.VFC<{
   gameId: number
@@ -426,7 +321,7 @@ const ViewOngoingGame: React.VFC<{
           'sm:p-3 sm:max-w-md sm:rounded-md sm:shadow-lg sm:border sm:border-gray-50'
         )}
       >
-        <ViewGamePlayers game={game} />
+        <PlayersRow game={game} />
 
         <div className="space-y-2">
           <DieRow.ViewWhite
